The Beginning of Time
This commit introduces the base functionality of the program. It
currently implements the following functions:

  punch (-h | -v)
  punch (in|out|log) [<project_name>]
  punch list

There is a known bug in this commit which occurs when you run `punch
log` before you ever run `punch in` or `punch out`. The bug occurs
because the program looks for `~/.punch/default.punch` and doesn't find
it because it doesn't exist yet. I plan to fix this in the near future.